unix: dossier2: fixing Solaris Qt3 support (bis)
- Casting pid_t to int (Solaris gcc warning)
- Erasing specified area on framePlan before update map
- Changing color Qt::transparent to Qt::color0
- Small bugs fixed on Solaris

+++ unix/QtDossierMetro2011/chemin/chemin.c
@@ -72,7 +72,7 @@ int main(void) {
 	AffChemin(&pathway);
 
 	/* Sending message */
-	debug("FRK: (DBG) Sending data to client (%d)\n", ask.client);
+	debug("FRK: (DBG) Sending data to client (%d)\n", (int) ask.client);
 	send_message(ACK_SEARCH, (void*) &pathway, ask.client, sizeof(pathway));
 	
 	exit(0);
